fix(type): fix all type conversion errors between Integer/Long and Boolean/Integer
- Change AssignRolesRequest.roleIds from List<Integer> to List<Long>
- Add Long to Integer conversions in UserServiceImpl for roleId mappings (roleId.intValue())
- Add Integer to Long conversion for existingRoleIds (getRoleId().longValue())
- Add Boolean to Integer conversions in MenuServiceImpl for visible field (? 1 : 0)
- Add Integer to Boolean conversions in MenuServiceImpl for visible field (== 1)
- Fix MenuController to call findByIdDto() instead of findById()
- Add Integer to Long conversion in RoleServiceImpl.convertMenuToResponse() for menu ID
- Add Boolean to Integer conversion in RoleServiceImpl for visible field

+++ src/main/java/org/example/jwtjavaeight/domain/dto/AssignRolesRequest.java
@@ -10,5 +10,5 @@ import java.util.List;
 @Setter
 public class AssignRolesRequest {
     @NotEmpty(message = "角色ID列表不能为空")
-    private List<Integer> roleIds;
+    private List<Long> roleIds;
 }
